backend/api/rule_engine/requirements.txt
cffi==1.16.0
clipspy==1.0.4
pycparser==2.22
